Problem 0035: Update dummy problem

+++ Numbers/SpecialNumbers/Primes/PrimeChecker.cs
@@ -7,7 +7,7 @@ public class PrimeChecker
     private readonly CachingEnumerator _primes = new(Prime.Create());
     private const long NoPrimeValue = -1;
 
-    public bool IsPrime(int candidate)
+    public bool IsPrime(long candidate)
     {
         var maybePrime = _primes.GetElements().SkipWhile(prime => prime < candidate).TakeWhile(prime => prime == candidate)
             .FirstOrDefault(NoPrimeValue);
